Add unit tests for the jService library
This adds unit tests, along with fixing one or two bugs that were revealed by
the tests. It does not implement tests for the unimplemented methods. As well,
it relies on the public api for expediency, but it would be better to mock the
part of the api we use so that we can guarantee data consistency.